+++ onecrm/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["__version__"]
 
-__version__ = "2.5.14"
+__version__ = "2.5.15"